pulled helper functions out into own file
